requirements.txt
RateMyProfessorPyAPI==1.2.1
<<<<<<< HEAD
=======
RateMyProfessorAPI==1.2.0
>>>>>>> 56441d84ec7a7471c1cfe84d1835f98301a12572
flask==0.12.2
discord.py==1.6.0
requests==2.18.4
matplotlib==3.3.2
bootstrap-flask==1.5.1